requirements.txt
requests
yt-dlp
Pillow
pyinstaller
